catalog-api/requirements.txt
asyncpg~=0.29.0
databases~=0.8.0
fastapi~=0.109.2
psycopg~=3.1.18
psycopg2-binary~=2.9.9
pydantic~=2.6.4
python-dotenv~=1.0.1
requests~=2.31.0
uvicorn~=0.29.0
